% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_count_device}
\alias{arrayfire_count_device}
\title{Return the number of available devices}
\usage{
arrayfire_count_device()
}
\value{
the number of available devices.
}
\description{
Return the number of available devices
}
\details{
Return the number of available devices.
}
\references{
\url{http://arrayfire.org/docs/group__device__func__count.htm#gac3c8750da69ee883fd14c0a8290e45f1}
}
