% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_device_info}
\alias{arrayfire_device_info}
\title{Report the information about current device and platform}
\usage{
arrayfire_device_info()
}
\value{
List containing the information about device and platform.
}
\description{
Report the information about current device and platform
}
\details{
Return a list containing the information about current device and platform.
The list object includes device name, palatform information, toolkit information,
and compute capability.
}
\references{
\url{http://arrayfire.org/docs/group__device__func__prop.htm#gaa9eab56ddac21650581682912ea78293}
}
