% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_device_prop}
\alias{arrayfire_device_prop}
\title{Report the information about current device and platform}
\usage{
arrayfire_device_prop()
}
\value{
List contains the information about device and platform.
}
\description{
Report the information about current device and platform
}
\details{
Return a list contains the informations about current device and platform.
The list object include device name, palatform information, toolkit information,
and compute capability.
}
\references{
\url{http://www.arrayfire.com/docs/group__device__mat.htm#gafae79277ec9ce681976c48392985baf0}
}

