% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_get_active_backend}
\alias{arrayfire_get_active_backend}
\title{Get string description for the active backend}
\usage{
arrayfire_get_active_backend()
}
\value{
string description for the active backend
}
\description{
Get string description for the active backend
}
\details{
Get string description for the active backend.
}
\references{
\url{http://arrayfire.org/docs/group__unified__func__getactivebackend.htm#gac6a8e976a151d007e0cf5cf4f11da2a9}
}
