% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_get_available_backends}
\alias{arrayfire_get_available_backends}
\title{Get a vector of string descriptions for the available backends}
\usage{
arrayfire_get_available_backends()
}
\value{
vector of string descriptions for the available backends
}
\description{
Get a vector of string descriptions for the available backends
}
\details{
Get a vector of string descriptions for the available backends.
}
\references{
\url{http://arrayfire.org/docs/group__unified__func__getavailbackends.htm#ga32fc807a8d4680987a915362573c1a2c}
}
