% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_get_device}
\alias{arrayfire_get_device}
\title{Return the index of current device}
\usage{
arrayfire_get_device()
}
\value{
the index of current device.
}
\description{
Return the index of current device
}
\details{
Get the index of current device which is lower or
equal to a value returned by \code{\link{arrayfire_count_device}}.
}
\references{
\url{http://arrayfire.org/docs/group__device__func__get.htm#ga4dfe3f90475b735384f8b28cf2b19a11}
}
