% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_get_seed}
\alias{arrayfire_get_seed}
\title{Get the seed for random number generation in ArrayFire.}
\usage{
arrayfire_get_seed()
}
\value{
a (64bit) double whose bit representation is reinterpreted from the 64bit integer seed
}
\description{
Get the seed for random number generation in ArrayFire.
}
\details{
Gets the seed for the current default random engine in ArrayFire. Note that a (64bit) double is used as output, since R does not support 64bit integers.
}
\references{
\url{http://arrayfire.org/docs/group__random__func__get__seed.htm}
}
