% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_info}
\alias{arrayfire_info}
\title{Display ArrayFire and device info}
\usage{
arrayfire_info(verbose = FALSE)
}
\arguments{
\item{verbose}{Flag to return verbose info}
}
\description{
Display ArrayFire and device info
}
\details{
Display ArrayFire and device info.
}
\references{
\url{http://arrayfire.org/docs/group__device__func__info.htm#ga1360e706b046b1c5b299bd6e689a0276} and \url{http://arrayfire.org/docs/group__device__func__info__string.htm#ga73ee21947f018445ea720036c1edfb69}
}
