% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_set_backend}
\alias{arrayfire_set_backend}
\title{Set the current backend}
\usage{
arrayfire_set_backend(backend = "DEFAULT")
}
\arguments{
\item{backend}{string description of backend (DEFAULT, CPU, OPENCL, CUDA)}
}
\description{
Set the current backend
}
\details{
Set the current backend.
}
\references{
\url{http://arrayfire.org/docs/group__unified__func__setbackend.htm#ga7e0fad1f134b0421811ee3a4ce47e987}
}
