% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_set_device}
\alias{arrayfire_set_device}
\title{Switch to specified device}
\usage{
arrayfire_set_device(index)
}
\arguments{
\item{index}{the index of a device switched to.}
}
\description{
Switch to specified device
}
\details{
Available indices
can be obtained by calling \code{\link{arrayfire_count_device}}.
}
\references{
\url{http://www.arrayfire.com/docs/group__device__mat.htm#ga6473f398cab654fc0bef97833a5bea5f}
}

