% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_set_seed}
\alias{arrayfire_set_seed}
\title{Set the seed for random number generation in ArrayFire.}
\usage{
arrayfire_set_seed(seed)
}
\arguments{
\item{seed}{a (64bit) double whose bit representation is reinterpreted as a 64bit integer}
}
\description{
Set the seed for random number generation in ArrayFire.
}
\details{
Sets the seed for the current default random engine in ArrayFire. Note that a (64bit) double is used as input, since R does not support 64bit integers.
}
\references{
\url{http://arrayfire.org/docs/group__random__func__set__seed.htm}
}
