% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastLmPure}
\alias{fastLmPure}
\title{Fast and simple linear model}
\usage{
fastLmPure(X, y)
}
\arguments{
\item{X}{a model matrix}

\item{y}{a vector containing the explained variable}
}
\value{
\code{fastLmPure} returns a list with three components:
\describe{
  \item{\code{coefficients}}{a vector of coefficients}
  \item{\code{stderr}}{a vector of the (estimated) standard errors of the coefficient estimates}
  \item{\code{df.residual}}{a scalar denoting the degrees of freedom in the model}
}
}
\description{
Fast and simple linear model
}
\details{
Single precision floats are used since not all devices support double precision.
}
