% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastRnorm}
\alias{fastRnorm}
\title{Fast rnorm alternative}
\usage{
fastRnorm(n, mean = 0, sd = 1, useDouble = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{mean}{mean value of the distribution}

\item{sd}{standard deviation of the distribution}

\item{useDouble}{use float or double internally}
}
\value{
A numeric vector of random values
}
\description{
Fast rnorm alternative
}
\details{
Fast generation of normaly distributed random values. The default is to use single precision,
 since not all devices support calculations using double precision.
}
\seealso{
\code{\link{rnorm}} and \url{http://arrayfire.org/docs/group__random__func__randn.htm}
}
