% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppArrayFire-package.R
\docType{package}
\name{RcppArrayFire-package}
\alias{RcppArrayFire}
\alias{RcppArrayFire-package}
\title{Rcpp integration for ArrayFire}
\description{
The RcppArrayFire package provides Rcpp integration for ArrayFire.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RInstitute/rcpparrayfire}
  \item \url{http://www.arrayfire.com}
  \item Report bugs at \url{https://github.com/RInstitute/rcpparrayfire/issues}
}

}
