% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_count_device}
\alias{arrayfire_count_device}
\title{Return the number of available devices}
\usage{
arrayfire_count_device()
}
\value{
the number of available devices.
}
\description{
Return the number of available devices
}
\details{
Return the number of available devices.
}
\references{
\url{http://www.arrayfire.com/docs/group__device__mat.htm#ga6d1964af8b32ca9c2f9525bf5ef8112a}
}

