% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_get_device}
\alias{arrayfire_get_device}
\title{Return the index of current device}
\usage{
arrayfire_get_device()
}
\value{
the index of current device.
}
\description{
Return the index of current device
}
\details{
Get the index of current device which is lower or
equal to a value returned by \code{\link{arrayfire_count_device}}.
}
\references{
\url{http://www.arrayfire.com/docs/group__device__mat.htm#ga6d1964af8b32ca9c2f9525bf5ef8112a}
}

