% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayfire_set_device}
\alias{arrayfire_set_device}
\title{Switch to specified device}
\usage{
arrayfire_set_device(index)
}
\arguments{
\item{index}{the index of a device switched to.}
}
\description{
Switch to specified device
}
\details{
Available indices
can be obtained by calling \code{\link{arrayfire_count_device}}.
}
\references{
\url{http://arrayfire.org/docs/group__device__func__set.htm#gafbb906ca5b89ec43fdb0e3a14d1df1e7}
}
